+++ source/extensions/omni.flux.commands/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.1]
+## Fixed
+- Fix deletion bug for empty references
+
 ## [1.2.0]
 ### Changed
 - Added `CreateOrInsertSublayerCommand` custom command for to revert Kit 106.5 changes.